refactor (limits): restructure feature limits module
