+++ src/modules/feature-limit/constants/platform.constants.ts
@@ -1,0 +1,10 @@
+export enum Platform {
+  WEB = 'web',
+  CLI = 'cli',
+  MOBILE = 'mobile',
+  DESKTOP = 'desktop',
+}
+
+export const PLATFORM_CONSTANTS = {
+  CLIENT_HEADER: 'internxt-client',
+} as const;
